refactor: InterruptibleInterval into a class
